Moves LiveQuery pub/sub to adapter folder
