+++ src/Adapters/PubSub/EventEmitterPubSub.js
