+++ src/LiveQuery/ParsePubSub.js
@@ -1,5 +1,11 @@
-import { RedisPubSub } from './RedisPubSub';
-import { EventEmitterPubSub } from './EventEmitterPubSub';
+import { loadAdapter } from '../Adapters/AdapterLoader';
+import {
+  EventEmitterPubSub
+} from '../Adapters/PubSub/EventEmitterPubSub';
+
+import {
+  RedisPubSub
+} from '../Adapters/PubSub/RedisPubSub';
 
 let ParsePubSub = {};
 
@@ -10,7 +16,7 @@ function useRedis(config: any): boolean {
 
 ParsePubSub.createPublisher = function(config: any): any {
   if (useRedis(config)) {
-    return RedisPubSub.createPublisher(config.redisURL);
+    return RedisPubSub.createPublisher(config);
   } else {
     return EventEmitterPubSub.createPublisher();
   }
@@ -18,7 +24,7 @@ ParsePubSub.createPublisher = function(config: any): any {
 
 ParsePubSub.createSubscriber = function(config: any): void {
   if (useRedis(config)) {
-    return RedisPubSub.createSubscriber(config.redisURL);
+    return RedisPubSub.createSubscriber(config);
   } else {
     return EventEmitterPubSub.createSubscriber();
   }
